picture_search/requirements.txt
cn_clip==1.5.1
numpy==2.1.3
onnxruntime_gpu==1.13.1
opensearch_py==2.7.1
paddleocr==2.9.1
paddleocr.egg==info
Pillow==11.0.0
torch==1.12.1+cu116
transformers==4.46.3
